+++ lovely/esdb/properties/__init__.py
@@ -1,2 +1,3 @@
 from .property import Property  # noqa
+from .relation import Relation  # noqa
 from .objectproperty import ObjectProperty  # noqa